+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.3
 
 - Segmented MethodChannel's in different channels to prevent overloading.
+- Fixed an issue on Android where the device would not disconnect properly. Now, when a disconnection is detected, the gatt will be closed.
 
 ## 1.2.2